refactor: moved regularity records to another file
